improve the registerScheduledTasksForTenant using spatie get tenant strategie
